Doing https://github.com/Shopify/shopify-app-template-remix/pull/1090 for JS branch
