+++ app/root.jsx
@@ -4,45 +4,16 @@ import {
   Outlet,
   Scripts,
   ScrollRestoration,
-  useLoaderData,
-  useNavigate,
 } from "@remix-run/react";
-import { authenticate } from "./shopify.server";
-import { useEffect } from "react";
-
-export const loader = async ({ request }) => {
-  await authenticate.admin(request);
-
-  return { apiKey: process.env.SHOPIFY_API_KEY || "" };
-};
 
 export default function App() {
-  const { apiKey } = useLoaderData();
-
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    const handleNavigate = (event) => {
-      const href = event.target.getAttribute("href");
-      if (href) navigate(href);
-    };
-
-    document.addEventListener("shopify:navigate", handleNavigate);
-
-    return () =>
-      document.removeEventListener("shopify:navigate", handleNavigate);
-  }, [navigate]);
-
   return (
     <html>
       <head>
         <meta charSet="utf-8" />
         <meta name="viewport" content="width=device-width,initial-scale=1" />
         <meta name="shopify-link-behavior" content="remix" />
-        <meta name="shopify-api-key" content={apiKey} />
         <link rel="preconnect" href="https://cdn.shopify.com/" />
-        <script src="https://cdn.shopify.com/shopifycloud/app-bridge.js"></script>
-        <script src="https://cdn.shopify.com/shopifycloud/app-bridge-ui-experimental.js"></script>
         <Meta />
         <Links />
       </head>
